Revise main page
