+++ src/components/SubMenu.tsx
@@ -1,0 +1,10 @@
+
+function SideMenu() {
+    return (
+        <>
+            <h2>Side menu (Placeholder, will be used to show selected patient data)</h2>
+        </>
+    )
+}
+
+export default SideMenu;
