+++ src/lib/Users.tsx
@@ -1,0 +1,8 @@
+const enum Users {
+    JHDoctor,
+    JHAdmin,
+    FDAAdmin,
+    BavariaAdmin
+};
+
+export default Users;
