+++ src/pages/index.tsx
@@ -1,267 +1,66 @@
 import React, { useState, useEffect, useReducer } from "react";
-import ResponsiveAppBar from '@/components/appbar';
-import { Typography,Button, Box, Stack, Modal, TextField, OutlinedInput,FormControl,InputAdornment, FormHelperText} from '@mui/material';
+import ResponsiveAppBar from '@/components/Appbar';
+import { Typography, Button, Box, Stack, Modal, TextField, OutlinedInput, FormControl, InputAdornment, FormHelperText, FormLabel, RadioGroup, FormControlLabel, Radio, Container, Drawer } from '@mui/material';
 import DataTable from '@/components/data-table';
-
-
-import useJaneHopkins from '../hooks/useJaneHopkins';
+import Users from "@/lib/Users";
+import AddPatientForm from "@/components/AddPatientForm";
+import SideMenu from "@/components/SubMenu";
 
 export default function Home() {
-  const { entities } = useJaneHopkins();
-  const [patients, setPatients] = React.useState<any[]>([]);
-  const [openViewModal, setOpenViewModal] = useState(false);//set to false so that it is closed
+  const [currentUser, setCurrentuser] = React.useState<Users>(Users.JHDoctor);
+  const onChange = (event: React.ChangeEvent<HTMLInputElement>, value: string) => {
 
-  
-
-  const handleCloseView = () => setOpenViewModal(false);
-
-//This stores the info that you inputted.
-  const [formInput, setFormInput] = useReducer(
-    (state: any, newState: any) => ({ ...state, ...newState }),
-    {
-      name: "",
-      dob: "",
-      height: "",
-      weight: "",
-      insuranceNumber: "",
-      address:"",
-
+    console.log("ASDASD");
+    switch (value) {
+      case "Users.JHDoctor":
+        setCurrentuser(Users.JHDoctor);
+        break;
+      case "Users.JHAdmin":
+        setCurrentuser(Users.JHAdmin);
+        break;
+      case "Users.FDAAdmin":
+        setCurrentuser(Users.FDAAdmin);
+        break;
+      case "Users.BavariaAdmin":
+        setCurrentuser(Users.BavariaAdmin);
+        break;
     }
-  );
-
-  //This creates the "name: formInput[0].name" to be set into the formInput that is above this.
-  const handleInput = (evt: any) => {
-    const name = evt.target.name;
-    const newValue = evt.target.value;
-    setFormInput({ [name]: newValue });
-    
-  };
-//Takes the information inputted into the formInput and stores it into the Vendia Database
-  const handleAdd = async () => {
-    const addPatientResponse = await entities.patient.add({
-      name:formInput.name,
-      dob:formInput.dob,
-      height:formInput.height,
-      weight:formInput.weight,
-      insuranceNumber:formInput.insuranceNumber,
-      address:formInput.address,
-      },
-      {
-        aclInput:{
-          acl:[
-            {
-              principal:{
-                nodes:["Bavaria", "FDA"]
-              },
-              operations: ["READ"], //Write gives the option to update
-              path:"weight",//What is allowed to go to Bavaria and FDA
-            },
-          ],
-        },
-      }
-    );
-    
-    setOpenViewModal(false); //Closes it
-  };
-  //This will reset the appearance of the form. Making blank once you enter it again.
-  //Probably a better way to do this.
-  const resetFormInput = async () =>{
-    console.log("reached here");
-    formInput.name = "";
-    formInput.dob = "";
-    formInput.height = "";
-    formInput.weight = "";
-    formInput.insuranceNumber = "";
-    formInput.address = "";
-  };
-
-  let rows: any[]= [];
-
-  const columns= [
-    { field: 'id', headerName: 'ID', width: 150 },
-    { field: 'name', headerName: 'Name', width: 150 },
-    { field: 'dob', headerName: 'Date of Birth', width: 175 },
-    { field: 'height', headerName: 'Height', width: 100 },
-    { field: 'weight', headerName: 'Weight', width: 100 },
-    { field: 'address', headerName: 'Address', width: 150 },
-    { field: 'insuranceNumber', headerName: 'Insurance Number', width: 300 },
-    { field: 'patientPicture', headerName: 'Patient Picture', width: 125 },
-    { field: 'bloodPressure', headerName: 'Blood Pressure', width: 100 },
-    { field: 'temperature', headerName: 'Temperature', width: 100 },
-    { field: 'oxygenSaturation', headerName: 'Oxygen Saturation', width: 100 },
-    { field: 'uuid', headerName: 'UUID', width: 100 },
-    { field: 'currentMedications', headerName: 'Current Medications', width: 100 },
-    { field: 'familyHistory', headerName: 'Family History', width: 100 },
-    { field: 'currentlyInsured', headerName: 'Currently Insured', width: 100 },
-    { field: 'icdHealthCodes', headerName: 'ICD Health Codes', width: 100 },
-    { field: 'allergies', headerName: 'Allergies', width: 100 },
-  ];
-
-
-
-
-  //List patients
-  const listPatients = async () => {
-    let patientList = await entities.patient.list();
-    setPatients(patientList.items);
-    
-  };
-  //Pushes out the info to the data table.
-  for (let i = 0; i < patients.length; i++) {
-  rows.push({
-    id: patients[i]._id,
-    name: patients[i].name,
-    dob: patients[i].dob,
-    height: patients[i].height,
-    weight: patients[i].weight,
-    address: patients[i].address,
-    insuranceNumber: patients[i].insuranceNumber,
-    patientPicture: patients[i].patientPicture,
-    bloodPressure: patients[i].bloodPressure,
-    temperature: patients[i].temperature,
-    oxygenSaturation: patients[i].oxygenSaturation,
-    uuid: patients[i].uuid,
-    currentMedications: patients[i].currentMedications,
-    familyHistory: patients[i].familyHistory,
-    currentlyInsured: patients[i].currentlyInsured,
-    icdHealthCodes: patients[i].icdHealthCodes,
-    allergies: patients[i].allergies,
-     }); 
-    }
-
-  useEffect(() => {
-    listPatients();
-    });
-
-   
+  }
 
   return (
     <>
       <ResponsiveAppBar />
-      <Button 
-          variant="contained"
-          onClick={() => {
-            resetFormInput();
-            setOpenViewModal(true);
-            
-          }}
-        >Add
-      </Button>
-      <DataTable
-        rows={rows}
-        columns={columns}
-        
-      />
-      <div>
-      <Modal
-        open={openViewModal}
-        onClose={handleCloseView}
-        aria-labelledby="modal-modal-title"
-        aria-describedby="modal-modal-description"
-        sx={{ zIndex: 1059 }}
-      >
-        <form onSubmit={handleAdd}>
-        <Box 
-          sx={{
-          position: "absolute",
-          top: "50%",
-          left: "50%",
-          transform: "translate(-50%, -50%)",
-          width: 400,
-          backgroundColor: "background.paper",
-          border: "2px solid #000",
-          boxShadow: 24,
-          p: 2,
-          }}
-        >
-          <Box sx={{
-            display: "flex",
-            justifyContent: "left",
-            alignItems: "center",
-            m:2,      
-            }}
-          >
-            <TextField sx = {{m: 1,width: 200 }} helperText="Name" name="name" onChange={handleInput} defaultValue={formInput.name} />
-            <TextField type="date" sx = {{m: 1,width: 350 }} helperText="Date of Birth" name="dob" onChange={handleInput} defaultValue={formInput.dob} />
-          </Box>
-          <Box sx={{
-            display: "flex",
-            justifyContent: "left",
-            alignItems: "center",
-            m:2,      
-            }}
-          >
-            <TextField sx = {{m: 1,width: 200 }} helperText="Address" name="address" onChange={handleInput}defaultValue={formInput.address} />
-          </Box>
-          <Box sx={{
-            display: "flex",
-            justifyContent: "left",
-            alignItems: "center",
-            m:2,      
-            }}
-          >
-            <TextField sx = {{m: 1,width: 200 }} helperText="Insurance" name="insuranceNumber" onChange={handleInput} defaultValue={formInput.insuranceNumber}/>
-          </Box>
-          <Box sx={{
-            display: "flex",
-            justifyContent: "left",
-            alignItems: "center",
-            m:2,      
-            }}
-          >
-            <FormControl sx={{ m: 1, width: '25ch' }} variant="outlined">
-            <OutlinedInput
-              id="outlined-adornment-height"
-              aria-describedby="outlined-height-helper-text"
-              inputProps={{
-                'aria-label': 'height',
-              }}
-              name="height"
-              onChange={handleInput}
-              defaultValue={formInput.height}
-            />
-            <FormHelperText id="outlined-height-helper-text">Height</FormHelperText>
-            </FormControl>
-            <FormControl sx={{ m: 1, width: '25ch' }} variant="outlined">
-            <OutlinedInput
-              id="outlined-adornment-weight"
-              endAdornment={<InputAdornment position="end">lbs</InputAdornment>}
-              aria-describedby="outlined-weight-helper-text"
-              inputProps={{
-                'aria-label': 'weight',
-              }}
-              name="weight"
-              onChange={handleInput}
-              defaultValue={formInput.weight}
-              
-            />
-            <FormHelperText id="outlined-weight-helper-text">Weight</FormHelperText>
-            </FormControl>
-            
-            </Box>
-        <Box>
-          <Button 
-            variant="contained"
-              onClick={handleAdd}
-            >Add Patient
-          </Button>
-            <Button 
-            variant="outlined"
-            onClick={() => {
-              setOpenViewModal(false);
-            }}
-          >Cancel
-          </Button></Box>
+      <br></br>
+      <Container  maxWidth={false} sx={{ display: 'flex', flexDirection: 'row', gap: '10px'}}>
+
+        <Box sx={{ border: 1, borderLeft: 0, borderColor: '#aaaaaa', borderTopRightRadius: '25px', borderBottomRightRadius: '25px', width: '20%', padding: '10px'}}>
+          <SideMenu/>
+          <FormControl>
+            <FormLabel id="form-user">User</FormLabel>
+            <RadioGroup
+              aria-labelledby="demo-radio-buttons-group-label"
+              defaultValue="Users.JHDoctor"
+              name="radio-buttons-group"
+              onChange={onChange}
+            >
+              <FormControlLabel value="Users.JHDoctor" control={<Radio />} label="JHDoctor" />
+              <FormControlLabel value="Users.JHAdmin" control={<Radio />} label="JHAdmin" />
+              <FormControlLabel value="Users.FDAAdmin" control={<Radio />} label="FDAAdmin" />
+              <FormControlLabel value="Users.BavariaAdmin" control={<Radio />} label="BavariaAdmin" />
+            </RadioGroup>
+          </FormControl>
         </Box>
-        </form>
-      </Modal>
-      
-    </div>
+
+        <Box sx={{flexGrow: 1}}>
+          <DataTable
+            currentUser={currentUser}
+          />
+          <AddPatientForm />
+        </Box>
+      </Container>
     </>
   );
 }
-
-
 
 /*
 //Allows for the changing of buttons from one to another.
@@ -326,4 +125,3 @@ export default function Home() {
       
       Shows stuff from the button
       */
-
